refactor: share member value helper
